+++ web/src/components/App.jsx
@@ -21,6 +21,7 @@ import Login from "./Login";
 import Signup from "./Signup";
 import Account from "./Account";
 import EmailVerify from "./EmailVerify";
+import PasswordReset from "./PasswordReset";
 import initI18n from "../app/i18n"; // Translations!
 import prefs from "../app/Prefs";
 import RTLCacheProvider from "./RTLCacheProvider";
@@ -65,6 +66,7 @@ const App = () => {
                   <Route path={routes.login} element={<Login />} />
                   <Route path={routes.signup} element={<Signup />} />
                   <Route path={routes.emailVerify} element={<EmailVerify />} />
+                  <Route path={routes.passwordReset} element={<PasswordReset />} />
                   <Route element={<Layout />}>
                     <Route path={routes.app} element={<AllSubscriptions />} />
                     <Route path={routes.account} element={<Account />} />